Updated Board Event Handling
Added TestBoard : to help debugging and model the incoming Puzzle Mode
Said board is initialised with test files named with the .pop extention 
 #Known Bugs: 
Chains do not increase more than 2. This is due to the fact that that the blocks are in a normal state 1 frame before the match, causing the chain to end there.